+++ packages/components/components/elvis-chip/src/react/declarations.d.ts
@@ -1,3 +1,2 @@
 declare module '*.css';
 declare module '*.scss';
-declare module '@elvia/elvis-colors';